Add video capturing module
This is a module for recording videos from the Half-Life engine.
Recording requires a runnable ffmpeg with libx264 for encoding.
Recording can be started with bxt_cap_start and stopped with
bxt_cap_stop.

This capturing module can be considered a successor to hl-capture [1].
It has a number of key architectural differences.

First, an external ffmpeg process is used rather than encoding inside
the HL1 process. This is useful mainly for lifting the 32-bit process
RAM limitation, allowing to capture 8K and above resolutions. It also
means I don't have to deal with the ffmpeg API.

Second, instead of OpenCL, Vulkan is used for converting colors from RGB
to YUV. Vulkan has much better driver support (surprisingly enough).
Notably, importing an OpenGL texture to OpenCL is not supported by Mesa,
which means that hl-capture's hardware acceleration was limited to
NVIDIA GPUs. Vulkan does not have this problem. Using Vulkan also comes
with a benefit of multi-threading support (not yet utilized, but
planned) and much better control to squeeze all possible performance.

Third, by the virtue of being built on top of the bxt-rs base, this
capturing module fully supports Windows.

[1]: https://github.com/YaLTeR/hl-capture