engine: Hook most Windows functions later
First we find the pointers, then BXT finds the pointers and hooks them
(changing the bytes), then we hook them too (changing the bytes again),
this way first our hook is called and then we call into BXT's hook, and
then BXT calls into hw.dll. This is the same order as what happens on
Linux.